Rename Base abstract class into Client
